requirements.txt
# Support Assistant Dependencies
# Python 3.9+

# HTTP клиент для API запросов
requests>=2.31.0

# Парсинг YAML конфигураций
PyYAML>=6.0.1

# Математические операции (косинусное сходство)
numpy>=1.24.0

# Красивый консольный вывод (опционально)
rich>=13.0.0

# Типизация (для Python < 3.10)
typing-extensions>=4.0.0
